requirements.txt
boto3
Click
jinja2
PyYaml
